Moved all XML code to private files
