+++ flutter_local_notifications_windows/lib/src/details/notification_action.dart
@@ -1,7 +1,5 @@
 import 'dart:io';
 
-import 'package:xml/xml.dart';
-
 // NOTE: All enum values in this file have Windows RT-specific names.
 // If you change their Dart names, be sure to override [Enum.name].
 
@@ -103,32 +101,4 @@ class WindowsAction {
 
   /// The tooltip, useful if [content] is empty.
   final String? tooltip;
-
-  /// Serializes this notification action as Windows-compatible XML.
-  ///
-  /// See: https://learn.microsoft.com/en-us/uwp/schemas/tiles/toastschema/element-action#syntax
-  void buildXml(XmlBuilder builder) {
-    if (image != null && !image!.isAbsolute) {
-      throw ArgumentError.value(
-        image!.path,
-        'WindowsImage.file',
-        'File path must be absolute',
-      );
-    }
-    builder.element(
-      'action',
-      attributes: <String, String>{
-        'content': content,
-        'arguments': arguments,
-        'activationType': activationType.name,
-        'afterActivationBehavior': activationBehavior.name,
-        if (placement != null) 'placement': placement!.name,
-        if (image != null) 'imageUri':
-          Uri.file(image!.absolute.path, windows: true).toFilePath(),
-        if (inputId != null) 'hint-inputId': inputId!,
-        if (buttonStyle != null) 'hint-buttonStyle': buttonStyle!.name,
-        if (tooltip != null) 'hint-toolTip': tooltip!,
-      },
-    );
-  }
 }
