+++ flutter_local_notifications_windows/lib/src/details/notification_image.dart
@@ -1,7 +1,5 @@
 import 'dart:io';
 
-import 'package:xml/xml.dart';
-
 import 'notification_part.dart';
 
 /// Where a Windows notification image can be placed.
@@ -44,25 +42,4 @@ class WindowsImage extends WindowsNotificationPart {
 
   /// How the image will be cropped. Null indicates uncropped.
   final WindowsImageCrop? crop;
-
-  @override
-  void buildXml(XmlBuilder builder) {
-    if (!file.isAbsolute) {
-      throw ArgumentError.value(
-        file.path,
-        'WindowsImage.file',
-        'File path must be absolute',
-      );
-    }
-    builder.element(
-      'image',
-      attributes: <String, String>{
-        'src': Uri.file(file.absolute.path, windows: true).toFilePath(),
-        'alt': altText,
-        'addImageQuery': addQueryParams.toString(),
-        if (placement != null) 'placement': placement!.name,
-        if (crop != null) 'hint-crop': crop!.name,
-      },
-    );
-  }
 }
