+++ flutter_local_notifications_windows/lib/src/details/xml/text.dart
@@ -1,0 +1,21 @@
+import 'package:xml/xml.dart';
+
+import '../notification_text.dart';
+
+/// Converts a [WindowsNotificationText] to XML
+extension TextToXml on WindowsNotificationText {
+  /// Serializes this text to Windows-compatible XML.
+  ///
+  /// See: https://learn.microsoft.com/en-us/uwp/schemas/tiles/toastschema/element-text
+  void buildXml(XmlBuilder builder) => builder.element(
+    'text',
+    attributes: <String, String>{
+      if (languageCode != null) 'lang': languageCode!,
+      if (placement != null) 'placement': placement!.name,
+      'hint-callScenarioCenterAlign': centerIfCall.toString(),
+      'hint-align': 'center',
+      if (isCaption) 'hint-style': 'captionsubtle',
+    },
+    nest: text,
+  );
+}
